Bind per-polygon debug data, and load NFS2 texture alignment data in for debug. Apply transformations to NFS2 polygons based upon findings. Cull NFS2 track properly, also cull from car position and not camera.

+++ src/Loaders/nfs3_loader.cpp
@@ -99,7 +99,7 @@ std::shared_ptr<TRACK> NFS3::LoadTrack(const std::string &track_base_path) {
     std::string track_name = p.filename().string();
     stringstream frd_path, col_path, can_path, hrz_path;
     string strip = "K0";
-    unsigned int pos = track_name.find(strip);
+    size_t pos = track_name.find(strip);
     if (pos != string::npos)
         track_name.replace(pos, strip.size(), "");
 
@@ -434,7 +434,7 @@ std::vector<TrackBlock> NFS3::ParseTRKModels(const std::shared_ptr<TRACK> &track
         // Get Verts from Trk block, indices from associated polygon block
         TRKBLOCK trk_block = track->trk[i];
         POLYGONBLOCK polygon_block = track->poly[i];
-        TrackBlock current_track_block(i,  rotationMatrix * glm::vec3(trk_block.ptCentre.x/ 10, trk_block.ptCentre.y/ 10, trk_block.ptCentre.z/ 10));
+        TrackBlock current_track_block(i, rotationMatrix * glm::vec3(trk_block.ptCentre.x/ 10, trk_block.ptCentre.y/ 10, trk_block.ptCentre.z/ 10));
         glm::vec3 trk_block_center = rotationMatrix * glm::vec3(0, 0, 0);
 
         // Light sources
